added test suite
